SSO: group related things in the same module #1567
Related symbols should live in the same module unless there is a clear
and conscious reason to separate them.
