+++ src/credentials/providers/sharedCredentialsProvider.ts
@@ -7,13 +7,13 @@ import * as AWS from 'aws-sdk'
 import { Profile } from '../../shared/credentials/credentialsFile'
 import { getLogger } from '../../shared/logger'
 import { getStringHash } from '../../shared/utilities/textUtilities'
-import { SSO_PROFILE_PROPERTIES, validateSsoProfile } from '../sso/ssoSupport'
 import { getMfaTokenFromUser } from '../credentialsCreator'
+import { hasProfileProperty } from '../credentialsUtilities'
+import { SSO_PROFILE_PROPERTIES, validateSsoProfile } from '../sso/sso'
+import { DiskCache } from '../sso/diskCache'
+import { SsoAccessTokenProvider } from '../sso/ssoAccessTokenProvider'
 import { CredentialsProvider } from './credentialsProvider'
 import { CredentialsProviderId } from './credentialsProviderId'
-import { hasProfileProperty } from '../credentialsUtilities'
-import { SsoAccessTokenProvider } from '../sso/ssoAccessTokenProvider'
-import { DiskCache } from '../sso/diskCache'
 import { SsoCredentialProvider } from './ssoCredentialProvider'
 
 const SHARED_CREDENTIAL_PROPERTIES = {
